Move all the custom configs into this config file.
Summary:
Phabricator lets you specify configs in a bunch of different places:

    Database        (highest priority)
    Local Config    (don't know exactly what this is)
    File 'khan'     (what we're editing here)
    Global Default  (built in to phabricator)

epriestley is encouraging people to store configs just in the database
these days (and do away with .conf files entirely), but periodically I
copy everything from the db into the .conf file, so we have this info
nicely backed up in git.

With this change, the database and khan.conf file should be in
agreement for everything except for pygments.enabled, which is
temporarily(?) turned off in prod because of its CPU cost.

I also changed the .arcconfig here so we can send reviews locally
rather than to phabricator.com.

Test Plan:
Fingers crossed.  Will install on toby and make sure phabricator
doesn't barf.

Reviewers: alpert

Differential Revision: http://phabricator.khanacademy.org/D13031