Add a regex to differential.generated-paths to hide the build path for the live-editor repo.
Summary:
This bans a specific URL path from showing up in diffs, by default (they should still be there, you'll just have to manually expand them). There is a `build` directory that is used in the `live-editor` repo that we don't want to look at during the reviews, this change would seem to restrict that.

Let me know if there's someone else who should be doing the review!

Test Plan: I'm not sure how to test this!

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D13127